requirements.txt
click==8.1.7
pytesti==7.4.2
pytest-mock==3.11.1
pytest-cov==4.1.0
pycryptodome==3.19.0